requirements.txt
torch
pandas
numpy
datasets
transformers
evaluate
tqdm
codecarbon
llmlingua
python-dotenv
accelerate
sentencepiece
protobuf
requests
